Too many reference images to choose from, so only the last 8 images are simply retained for selection.

+++ agents/video_generation_pipeline.py
@@ -62,20 +62,23 @@ class VideoGenerationPipeline:
         character_text_dir = os.path.join(self.working_dir, "text", "characters")
     
         # extract and save the characters
-        characters = self.storyboard_generator.extract_characters(script)
-        os.makedirs(character_text_dir, exist_ok=True)
-        for character in characters:
-            with open(f"{character_text_dir}/character_{character.idx}.json", "w") as f:
-                json.dump(character.model_dump(), f, indent=4, ensure_ascii=False)
+        if os.path.exists(character_text_dir) and len(os.listdir(character_text_dir)) > 0:
+            # load the characters
+            characters = []
+            for file_name in os.listdir(character_text_dir):
+                if file_name.endswith(".json"):
+                    with open(os.path.join(character_text_dir, file_name), "r") as f:
+                        character_dict = json.load(f)
+                        character: Character = Character.model_validate(character_dict)
+                        characters.append(character)
+        else:
+            characters = self.storyboard_generator.extract_characters(script)
+            os.makedirs(character_text_dir, exist_ok=True)
+            for character in characters:
+                with open(f"{character_text_dir}/character_{character.idx}.json", "w") as f:
+                    json.dump(character.model_dump(), f, indent=4, ensure_ascii=False)
 
-        # load the characters
-        characters = []
-        for file_name in os.listdir(character_text_dir):
-            if file_name.endswith(".json"):
-                with open(os.path.join(character_text_dir, file_name), "r") as f:
-                    character_dict = json.load(f)
-                    character: Character = Character.model_validate(character_dict)
-                    characters.append(character)
+
 
         # 2. generate the character portrait
         character_image_dir = os.path.join(self.working_dir, "images", "characters")
@@ -85,7 +88,8 @@ class VideoGenerationPipeline:
             save_dir = os.path.join(character_image_dir, character.identifier)
             logging.info(f"Generating portrait for character: {character.identifier} (save to {save_dir})")
             os.makedirs(save_dir, exist_ok=True)
-            self.character_generator(character, style, save_dir)
+            if not os.path.exists(os.path.join(save_dir, "front.png")):
+                self.character_generator(character, style, save_dir)
 
             for image_name in os.listdir(save_dir):
                 view_image_path = os.path.join(save_dir, image_name)
@@ -101,11 +105,14 @@ class VideoGenerationPipeline:
         os.makedirs(shot_image_dir, exist_ok=True)
         while True:
             # 3.1 design next shot
-            shot = self.storyboard_generator.get_next_shot_description(script, characters, shots)
 
-            # save the shot
-            with open(f"{shot_text_dir}/shot_{shot.idx}.json", "w") as f:
-                json.dump(shot.model_dump(), f, indent=4, ensure_ascii=False)
+            shot_text_path = f"{shot_text_dir}/shot_{len(shots)}.json"
+            if not os.path.exists(shot_text_path):
+                shot = self.storyboard_generator.get_next_shot_description(script, characters, shots)
+
+                # save the shot
+                with open(f"{shot_text_dir}/shot_{shot.idx}.json", "w") as f:
+                    json.dump(shot.model_dump(), f, indent=4, ensure_ascii=False)
 
             # load the shot
             with open(f"{shot_text_dir}/shot_{len(shots)}.json", "r") as f:
@@ -122,29 +129,31 @@ class VideoGenerationPipeline:
                     logging.info(f"Shot {shot.idx} does not have {frame_type} attribute")
                     continue
 
-                # 3.2.1 select refence image and generate guidance prompt
-                logging.info(f"Selecting reference images for shot {shot.idx} - {frame_type}")
-                ref_image_indices_and_text_prompt = self.reference_image_selector(available_image_path_and_text_pairs, getattr(shot, frame_type))
-                ref_image_path_and_text_pairs = [
-                    available_image_path_and_text_pairs[i]
-                    for i in ref_image_indices_and_text_prompt.ref_image_indices
-                ]
-                guide_prompt = ref_image_indices_and_text_prompt.text_prompt
-
-                # generate candidate images, then select the best one
-                logging.info(f"Reference images included for shot {shot.idx} - {frame_type}: \n{ref_image_path_and_text_pairs}")
-                logging.info(f"Generating {frame_type} for shot {shot.idx}: \n{guide_prompt}")
-                candidate_save_dir = os.path.join(shot_image_dir, f"shot_{shot.idx}-{frame_type}_candidate")
-                os.makedirs(candidate_save_dir, exist_ok=True)
+                # 3.2.1 select reference image and generate guidance prompt
                 best_save_path = os.path.join(shot_image_dir, f"shot_{shot.idx}-{frame_type}.png")
+                if not os.path.exists(best_save_path):
+                    logging.info(f"Selecting reference images for shot {shot.idx} - {frame_type}")
+                    print(available_image_path_and_text_pairs)
+                    ref_image_indices_and_text_prompt = self.reference_image_selector(available_image_path_and_text_pairs, getattr(shot, frame_type))
+                    ref_image_path_and_text_pairs = [
+                        available_image_path_and_text_pairs[i]
+                        for i in ref_image_indices_and_text_prompt.ref_image_indices
+                    ]
+                    guide_prompt = ref_image_indices_and_text_prompt.text_prompt
 
-                self.frame_image_generator(
-                    ref_image_path_and_text_pairs=ref_image_path_and_text_pairs,
-                    guide_prompt=guide_prompt,
-                    frame_description=getattr(shot, frame_type),
-                    candidate_save_dir=candidate_save_dir,
-                    best_save_path=best_save_path,
-                )
+                    # generate candidate images, then select the best one
+                    logging.info(f"Reference images included for shot {shot.idx} - {frame_type}: \n{ref_image_path_and_text_pairs}")
+                    logging.info(f"Generating {frame_type} for shot {shot.idx}: \n{guide_prompt}")
+                    candidate_save_dir = os.path.join(shot_image_dir, f"shot_{shot.idx}-{frame_type}_candidate")
+                    os.makedirs(candidate_save_dir, exist_ok=True)
+
+                    self.frame_image_generator(
+                        ref_image_path_and_text_pairs=ref_image_path_and_text_pairs,
+                        guide_prompt=guide_prompt,
+                        frame_description=getattr(shot, frame_type),
+                        candidate_save_dir=candidate_save_dir,
+                        best_save_path=best_save_path,
+                    )
 
                 # while True:
                 #     # 3.2.2 generate the frame image
@@ -174,15 +183,20 @@ class VideoGenerationPipeline:
                 frame_paths.append(best_save_path)
                 available_image_path_and_text_pairs.append((best_save_path, getattr(shot, frame_type)))
 
+                # TODO 这里只简单地选择了最后面8张图像，后面可以改成通过文本筛选
+                if len(available_image_path_and_text_pairs) > 8:
+                    available_image_path_and_text_pairs = available_image_path_and_text_pairs[-8:]
+
 
             # 4. generate video
             os.makedirs(os.path.join(self.working_dir, "videos"), exist_ok=True)
             video_save_path = os.path.join(self.working_dir, "videos", f"shot_{shot.idx}.mp4")
-            self.video_generator.generate_video(
-                prompt=shot.visual_content,
-                image_paths=frame_paths,
-                save_path=video_save_path,
-            )
+            if not os.path.exists(video_save_path):
+                self.video_generator.generate_video(
+                    prompt=shot.visual_content,
+                    image_paths=frame_paths,
+                    save_path=video_save_path,
+                )
 
             if shot.is_last:
                 break
